DevBranchesTest test restrictions

+++ tests/Common/DevBranchesTest.php
@@ -11,32 +11,66 @@ use Keboola\Test\StorageApiTestCase;
 
 class DevBranchesTest extends StorageApiTestCase
 {
-    private static $cleanupAfterClassTokenId;
-
-    private static $teardownClient;
-
-    /**
-     * @dataProvider provideValidClients
-     */
-    public function testCreateBranch(Client $providedClient)
+    public function setUp()
     {
-        $providedToken = $providedClient->verifyToken();
-        $branches = new DevBranches($providedClient);
+        parent::setUp();
 
-        // cleanup
-        $branchesList = $branches->listBranches();
-        $branchName = __CLASS__ . '\\' . $this->getName() . '\\' . $providedToken['id'];
-        $branchesCreatedByThisTestMethod = array_filter(
-            $branchesList,
-            function ($branch) use ($branchName) {
-                return strpos($branch['name'], $branchName) === 0;
+        // branches cleanup
+        $branches = new DevBranches($this->_client);
+        foreach ($branches->listBranches() as $branch) {
+            if ($branch['isDefault'] !== true) {
+                $branches->deleteBranch($branch['id']);
             }
-        );
-        foreach ($branchesCreatedByThisTestMethod as $branch) {
-            $branches->deleteBranch($branch['id']);
         }
+    }
 
-        // test
+    public function testCannotDeleteDefaultBranch()
+    {
+        $branches = new DevBranches($this->_client);
+        $branchesList = $branches->listBranches();
+
+        $this->assertCount(1, $branchesList);
+        $branch = reset($branchesList);
+
+        $this->assertSame(true, $branch['isDefault']);
+
+        $this->expectException(ClientException::class);
+        $this->expectExceptionCode(400);
+        $this->expectExceptionMessage('Cannot delete Main branch');
+
+        $branches->deleteBranch($branch['id']);
+
+        $this->assertCount(1, $branches->listBranches());
+    }
+
+    public function testAdminRoleBranchesPermissions()
+    {
+        $description = __CLASS__ . '\\' . $this->getName();
+
+        $adminDevBranches = new DevBranches($this->_client);
+        $branch = $adminDevBranches->createBranch($description);
+
+
+        $branches = new DevBranches($this->getClientForToken(STORAGE_API_GUEST_TOKEN));
+        $guestBranch = $branches->createBranch($description . '\\GuestRole');
+
+        $this->assertCount(3, $adminDevBranches->listBranches());
+
+        $this->assertSame($branch, $adminDevBranches->getBranch($branch['id']));
+        $this->assertSame($guestBranch, $adminDevBranches->getBranch($guestBranch['id']));
+
+        $adminDevBranches->deleteBranch($branch['id']);
+        $adminDevBranches->deleteBranch($guestBranch['id']);
+
+        $this->assertCount(1, $adminDevBranches->listBranches());
+    }
+
+    public function testBranchCreateAndDelete()
+    {
+        $token = $this->_client->verifyToken();
+        $branches = new DevBranches($this->_client);
+
+        $branchName = __CLASS__ . '\\' . $this->getName();
 
         // can create branch
         $branch = $branches->createBranch($branchName);
@@ -47,14 +81,15 @@ class DevBranchesTest extends StorageApiTestCase
         $this->assertArrayNotHasKey('admin', $branch);
         $this->assertArrayHasKey('creatorToken', $branch);
         $this->assertArrayHasKey('id', $branch['creatorToken']);
-        $this->assertEquals($providedToken['id'], $branch['creatorToken']['id']);
+        $this->assertEquals($token['id'], $branch['creatorToken']['id']);
         $this->assertArrayHasKey('name', $branch['creatorToken']);
-        $this->assertSame($providedToken['description'], $branch['creatorToken']['name']);
+        $this->assertSame($token['description'], $branch['creatorToken']['name']);
         $this->assertSame($branchName, $branch['name']);
+        $this->assertSame(false, $branch['isDefault']);
         $branchId = $branch['id'];
 
         // event is created for created branch
-        $event = $this->findLastEvent($providedClient, [
+        $event = $this->findLastEvent($this->_client, [
             'event' => 'storage.devBranchCreated',
             'objectId' => $branchId,
         ]);
@@ -70,14 +105,15 @@ class DevBranchesTest extends StorageApiTestCase
         $this->assertArrayNotHasKey('admin', $branchFromDetail);
         $this->assertArrayHasKey('creatorToken', $branchFromDetail);
         $this->assertArrayHasKey('id', $branchFromDetail['creatorToken']);
-        $this->assertEquals($providedToken['id'], $branchFromDetail['creatorToken']['id']);
+        $this->assertEquals($token['id'], $branchFromDetail['creatorToken']['id']);
         $this->assertArrayHasKey('name', $branchFromDetail['creatorToken']);
-        $this->assertSame($providedToken['description'], $branchFromDetail['creatorToken']['name']);
+        $this->assertSame($token['description'], $branchFromDetail['creatorToken']['name']);
+        $this->assertSame(false, $branch['isDefault']);
 
         // can list branches and see created branch
-        $branchList = $branches->listBranches();
-        $this->assertGreaterThanOrEqual(1, count($branchList));
-        $this->assertContains($branchFromDetail, $branchList);
+        $branchesList = $branches->listBranches();
+        $this->assertCount(2, $branchesList);
+        $this->assertContains($branchFromDetail, $branchesList);
 
         $defaultBranchCount = 0;
         foreach ($branchesList as $branch) {
@@ -93,6 +129,7 @@ class DevBranchesTest extends StorageApiTestCase
             $branches->createBranch($branchName);
             $this->fail('Sharing bucket to non organization member should fail.');
         } catch (ClientException $e) {
+            $this->assertSame(400, $e->getCode());
             $this->assertSame(
                 sprintf('There already is a branch with name "%s"', $branchName),
                 $e->getMessage()
@@ -104,7 +141,7 @@ class DevBranchesTest extends StorageApiTestCase
         $branches->deleteBranch($branchId);
 
         // there is event for deleted branch
-        $this->findLastEvent($providedClient, [
+        $this->findLastEvent($this->_client, [
             'event' => 'storage.devBranchDeleted',
             'objectId' => $branchId,
         ]);
@@ -113,99 +150,120 @@ class DevBranchesTest extends StorageApiTestCase
         try {
             $branches->deleteBranch($branchId);
         } catch (ClientException $e) {
-            $this->assertSame(
-                sprintf('Branch not found'),
-                $e->getMessage()
-            );
+            $this->assertSame(404, $e->getCode());
+            $this->assertSame(sprintf('Branch not found'), $e->getMessage());
         }
 
         // now branch can be created with same name as deleted branch
-        $newBranch = $branches->createBranch($branchName);
-    }
-
-    public function testOrgAdminCanDeleteBranchCreatedByAdmin()
-    {
-        $guestClient = $this->getGuestStorageApiClient();
-
-        $branches = new DevBranches($guestClient);
-        $branchName = __CLASS__ . ' příliš žluťoučký kůň' . microtime();
-
-        $branch = $branches->createBranch($branchName);
-
-        $branchId = $branch['id'];
-
-        $orgAdminClient = $this->getDefaultClient();
-        $branches = new DevBranches($orgAdminClient);
-
-        $branches->deleteBranch($branchId);
-
-        $this->expectNotToPerformAssertions();
-    }
-
-    public function testGuestCannotDeleteNotOwnedBranch()
-    {
-        $orgAdminClient = $this->getDefaultClient();
-        $branches = new DevBranches($orgAdminClient);
-
-        $branchName = __CLASS__ . ' příliš žluťoučký kůň' . microtime();
-
-        $branch = $branches->createBranch($branchName);
-
-        $branchId = $branch['id'];
-
-        $guestClient = $this->getGuestStorageApiClient();
-        $branches = new DevBranches($guestClient);
-
-        $this->expectException(ClientException::class);
-        $this->expectExceptionMessage('You don\'t have access to resource.');
-
-        $branches->deleteBranch($branchId);
-    }
-
-    public function provideValidClients()
-    {
-        $guest = $this->getGuestStorageApiClient();
-        return [
-            'admin' => [$this->getDefaultClient()],
-            'guest' => [$guest],
-        ];
-    }
-
-    /**
-     * @dataProvider provideInvalidClients
-     */
-    public function testInsufficientClientCannotCreateBranch(Client $providedClient)
-    {
-        $branches = new DevBranches($providedClient);
-
-        $branchName = __CLASS__ . time();
-
-        $this->expectExceptionMessage('You don\'t have access to resource.');
-        $this->expectException(ClientException::class);
-
         $branches->createBranch($branchName);
     }
 
-    public function provideInvalidClients()
+
+    public function testGuestRoleBranchesPermissions()
     {
-        self::$teardownClient = $this->getDefaultClient();
-        $tokenId = self::$teardownClient->createToken(new TokenCreateOptions());
-        $token = self::$teardownClient->getToken($tokenId);
-        self::$cleanupAfterClassTokenId = $token['id'];
+        $description = __CLASS__ . '\\' . $this->getName();
 
-        $notAdminClient = $this->getClientForToken($token['token']);
+        $adminDevBranches = new DevBranches($this->_client);
+        $branch = $adminDevBranches->createBranch($description);
 
-        return [
-            'not admin' => [$notAdminClient],
-        ];
+        $this->assertCount(2, $adminDevBranches->listBranches());
+
+        $branches = new DevBranches($this->getClientForToken(STORAGE_API_GUEST_TOKEN));
+        $branches->createBranch($description . '\\GuestRole');
+
+        $this->assertCount(3, $branches->listBranches());
+
+        $this->assertSame($branch, $branches->getBranch($branch['id']));
+
+        try {
+            $branches->deleteBranch($branch['id']);
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
+        }
+
+        $this->assertCount(3, $adminDevBranches->listBranches());
     }
 
-    public static function tearDownAfterClass()
+    public function testNonAdminTokenRestrictions()
     {
-        parent::tearDownAfterClass();
-        if (self::$cleanupAfterClassTokenId) {
-            self::$teardownClient->dropToken(self::$cleanupAfterClassTokenId);
+        $description = __CLASS__ . '\\' . $this->getName();
+
+        $adminDevBranches = new DevBranches($this->_client);
+        $branch = $adminDevBranches->createBranch($description);
+
+        $options = (new TokenCreateOptions())
+            ->setDescription($description)
+            ->setCanManageBuckets(true)
+            ->setExpiresIn(3600)
+        ;
+
+        $tokenId = $this->_client->createToken($options);
+        $token = $this->_client->getToken($tokenId);
+
+        $branches = new DevBranches($this->getClientForToken($token['token']));
+
+        try {
+            $branches->createBranch($description);
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
         }
-        self::$teardownClient = null;
+
+        $this->assertCount(2, $adminDevBranches->listBranches());
+
+        try {
+            $branches->getBranch($branch['id']);
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
+        }
+
+        try {
+            $branches->listBranches();
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
+        }
+
+        try {
+            $branches->deleteBranch($branch['id']);
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
+        }
+
+        $this->assertCount(2, $adminDevBranches->listBranches());
+    }
+
+    public function testReadOnlyRoleBranchesPermissions()
+    {
+        $description = __CLASS__ . '\\' . $this->getName();
+
+        $adminDevBranches = new DevBranches($this->_client);
+        $branch = $adminDevBranches->createBranch($description);
+
+        $this->assertCount(2, $adminDevBranches->listBranches());
+
+        $branches = new DevBranches($this->getClientForToken(STORAGE_API_READ_ONLY_TOKEN));
+
+        try {
+            $branches->createBranch($description);
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
+        }
+
+        $this->assertCount(2, $branches->listBranches());
+
+        $this->assertSame($branch, $branches->getBranch($branch['id']));
+
+        try {
+            $branches->deleteBranch($branch['id']);
+        } catch (ClientException $e) {
+            $this->validateAccessForbiddenException($e);
+        }
+
+        $this->assertCount(2, $adminDevBranches->listBranches());
+    }
+
+    private function validateAccessForbiddenException(ClientException $exception)
+    {
+        $this->assertSame(403, $exception->getCode());
+        $this->assertSame('You don\'t have access to resource.', $exception->getMessage());
     }
 }
